chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai>=1.0.0
pyjwt==2.9.0
httpx==0.27.2
pydantic-settings==2.5.2
